#554: move storage.cc/h to src/storage
