+++ src/cds_objects.cc
@@ -31,7 +31,7 @@
 
 #include "cds_objects.h"
 #include "mxml/mxml.h"
-#include "storage.h"
+#include "storage/storage.h"
 #include "util/tools.h"
 
 using namespace zmm;
