+++ src/search_handler.h
@@ -30,7 +30,7 @@
 
 #include "cds_objects.h"
 #include "memory.h"
-#include "storage.h"
+#include "storage/storage.h"
 #include "zmm/ref.h"
 
 #include <memory>
