+++ src/update_manager.cc
@@ -33,7 +33,7 @@
 
 #include "server.h"
 #include "singleton.h"
-#include "storage.h"
+#include "storage/storage.h"
 #include "util/tools.h"
 #include "upnp_cds.h"
 #include <chrono>
