+++ src/web/edit_load.cc
@@ -33,7 +33,7 @@
 
 #include "cds_objects.h"
 #include "common.h"
-#include "storage.h"
+#include "storage/storage.h"
 #include "util/tools.h"
 #include <cstdio>
