+++ src/web/edit_save.cc
@@ -32,7 +32,7 @@
 #include "content_manager.h"
 #include "zmm/dictionary.h"
 #include "pages.h"
-#include "storage.h"
+#include "storage/storage.h"
 #include "util/tools.h"
 #include <cstdio>
