+++ src/web/items.cc
@@ -32,7 +32,7 @@
 #include "cds_objects.h"
 #include "common.h"
 #include "pages.h"
-#include "storage.h"
+#include "storage/storage.h"
 #include "upnp_xml.h"
 
 using namespace zmm;
