+++ src/web/web_update.cc
@@ -32,7 +32,7 @@
 #include "autoscan.h"
 #include "content_manager.h"
 #include "pages.h"
-#include "storage.h"
+#include "storage/storage.h"
 
 using namespace zmm;
 using namespace mxml;
